endrer på padding for skatt, og gjør navytelser lik som skatt ihht zeplin skisser
