+++ web/src/frontend/src/digisos/skjema/inntektFormue/skattbarInntekt/inntektTypes.ts
