+++ web/src/frontend/src/digisos/skjema/inntektFormue/utbetalinger/utbetalingerTypes.tsx
